# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Application API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the generated application",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://your-production-url.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/expense-report/{id}": {
      "get": {
        "summary": "Get expense report by ID",
        "description": "Retrieve a specific expense report record",
        "tags": [
          "Expense Report"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved expense report"
          },
          "404": {
            "description": "Expense report not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update expense report",
        "description": "Update an existing expense report record",
        "tags": [
          "Expense Report"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated"
          },
          "404": {
            "description": "Expense report not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete expense report",
        "description": "Delete an expense report record",
        "tags": [
          "Expense Report"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted"
          },
          "404": {
            "description": "Expense report not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/expense-report": {
      "get": {
        "summary": "Get all expense reports",
        "description": "Retrieve all expense report records with pagination",
        "tags": [
          "Expense Report"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 100
            },
            "description": "Maximum number of records to return"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Number of records to skip"
          },
          {
            "in": "query",
            "name": "employee_id",
            "schema": {
              "type": "string"
            },
            "description": "Filter by employee ID"
          },
          {
            "in": "query",
            "name": "department",
            "schema": {
              "type": "string"
            },
            "description": "Filter by department"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved expense reports"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new expense report",
        "description": "Create a new expense report record",
        "tags": [
          "Expense Report"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "employee_name",
                  "employee_id",
                  "reimbursement_address",
                  "phone",
                  "department",
                  "manager",
                  "manager_phone",
                  "destination",
                  "purpose",
                  "from_date",
                  "to_date",
                  "expense_date",
                  "description"
                ],
                "properties": {
                  "employee_name": {
                    "type": "string"
                  },
                  "employee_id": {
                    "type": "string"
                  },
                  "reimbursement_address": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "department": {
                    "type": "string"
                  },
                  "manager": {
                    "type": "string"
                  },
                  "manager_phone": {
                    "type": "string"
                  },
                  "destination": {
                    "type": "string"
                  },
                  "purpose": {
                    "type": "string"
                  },
                  "from_date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "to_date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "expense_date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "description": {
                    "type": "string"
                  },
                  "air_and_transport": {
                    "type": "number"
                  },
                  "lodging": {
                    "type": "number"
                  },
                  "fuel": {
                    "type": "number"
                  },
                  "phone_expenses": {
                    "type": "number"
                  },
                  "meals_and_tips": {
                    "type": "number"
                  },
                  "entertainment": {
                    "type": "number"
                  },
                  "other_expenses": {
                    "type": "number"
                  },
                  "total": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Expense report created successfully"
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}